backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.2.0
pydantic==2.5.2
boto3==1.34.0
botocore==1.34.0
httpx==0.25.2
python-dotenv==1.0.0
requests==2.31.0 